77 move functionality to pipelines (#106)
