+++ src/cnn/model_api.py
@@ -5,9 +5,9 @@ Module for classifying audio data using a CNN model.
 
 """
 
-from src.audio.wav import FlattenWavIterator, AudioDataIterator
 from src.cnn.cnn import BasicCNN
 from src.constants import MODEL_WINDOW_LENGTH, WavIteratorType, CLASSIFICATION_CONFIDENCE_THRESHOLD
+from src.pipeline.wav import FlattenWavIterator, AudioDataIterator
 
 
 def classify_file(file_path: str, model: BasicCNN) -> bool:
