+++ src/frontend/app_state.py
@@ -20,7 +20,7 @@ class AppState:
     page: str
     classifier: BasicCNN
 
-    def __init__(self):
+    def __init__(self) -> None:
         """
         Initialize the application state
         """
