--- src/helper_scripts/work_dir/.gitignore
@@ -1,1 +1,0 @@
-*
