+++ src/pipeline/classifier.py
@@ -5,8 +5,8 @@ This module contains the Classifier class, which provides functionality
 for fitting a model and making predictions based on audio data.
 """
 
-import numpy as np
-
+import torch
+from torch import nn
 
 class Classifier:
     """
@@ -16,19 +16,22 @@ class Classifier:
     to predict labels for new data inputs.
     """
 
-    def __init__(self):
+    def __init__(self, model: nn.Module) -> None:
         """
         Initializes the Classifier instance.
         """
-        return
+
+        self.model = model
+        self.device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')
+        self.model.to(self.device)
 
     # pylint: disable=unused-argument
-    def fit(self, x_data: list[np.ndarray], y_data: list[int]):
+    def fit(self, x_data: list[torch.Tensor], y_data: list[int]) -> 'Classifier':
         """
         Fit the classifier to the training data.
 
         Args:
-            x_data (list[np.ndarray]): A list of input data arrays (features).
+            x_data (list[torch.Tensor]): A list of input data arrays (features).
             y_data (list[int]): A list of labels corresponding to the input data.
 
         Returns:
@@ -36,17 +39,26 @@ class Classifier:
         """
         return self
 
-    def predict(self, x_data: list[np.ndarray]) -> np.ndarray:
+    def predict(self, x_data: list[torch.Tensor]) -> list[int]:
         """
         Predict labels for the given input data.
 
         Args:
-            x_data (list[np.ndarray]): A list of input data arrays (features) 
+            x_data (list[torch.Tensor]): A list of input data arrays (features) 
             to predict labels for.
 
         Returns:
             list[int]: A list of predicted labels.
         """
-        # Implement prediction logic here
-        # For example, return a list of predicted labels
-        return [0 for _ in x_data]
+
+        predictions = []
+
+        with torch.no_grad():
+            for tens in x_data:
+                tens = tens.to(self.device)
+
+                prediction = self.model(tens)
+                predicted_label = prediction.argmax(0).item()
+                predictions.append(predicted_label)
+
+        return predictions
